[Testing] consolidate benchmarking code
