+++ CODEOWNERS
@@ -38,9 +38,7 @@
 
 # Performance Stream
 /integration/benchmark/** @SaveTheRbtz @simonhf @Kay-Zee @pattyshack @AndrewM-SDET
-/integration/loader/** @SaveTheRbtz @simonhf @Kay-Zee @pattyshack @AndrewM-SDET
 /integration/localnet/** @SaveTheRbtz @simonhf @Kay-Zee @zhangchiqing @pattyshack @AndrewM-SDET
-/integration/utils/** @SaveTheRbtz @simonhf @Kay-Zee @pattyshack @AndrewM-SDET
 /module/profiler/** @SaveTheRbtz @simonhf @Kay-Zee @pattyshack
 /module/trace/** @SaveTheRbtz @simonhf @Kay-Zee @pattyshack
 /module/tracer.go @SaveTheRbtz @simonhf @Kay-Zee @pattyshack
