+++ integration/Makefile
@@ -8,14 +8,6 @@ integration-test: access-tests ghost-tests mvp-tests execution-tests verificatio
 .PHONY: ci-integration-test
 ci-integration-test: access-tests ghost-tests mvp-tests epochs-tests consensus-tests execution-tests verification-tests network-tests collection-tests
 
-.PHONY: benchmark
-benchmark:
-	go test -v -tags relic -count=1 -timeout 30m ./benchmark -run Benchmark
-
-.PHONY: ci-benchmark
-ci-benchmark:
-	ENV=TEAMCITY go test -v -tags relic -count=1 -timeout 15m ./benchmark -run Benchmark
-
 ############################################################################################
 # CAUTION: DO NOT MODIFY THE TARGETS BELOW! DOING SO WILL BREAK THE FLAKY TEST MONITOR
 # In particular, do not skip tests by commenting them out here.
