+++ integration/benchmark/cmd/manual/Dockerfile
