+++ integration/benchmark/scripts/myFavContract.cdc
