+++ integration/benchmark/worker_stats_tracker.go
@@ -1,4 +1,4 @@
-package utils
+package benchmark
 
 import (
 	"fmt"
