+++ integration/localnet/Makefile
@@ -74,7 +74,7 @@ stop:
 
 .PHONY: load
 load:
-	go run --tags relic ../loader -log-level info -tps 1,0,10,0,100 -tps-durations 30s,10s,30s,10s
+	go run --tags relic ../benchmark/cmd/manual -log-level info -tps 1,0,10,0,100 -tps-durations 30s,10s,30s,10s
 
 .PHONY: tps-test
 tps-test:
